Added some checks before deleting files that the previous command returned a 0.  For the _dir files it seems to me that the comparative operators for the return values were wrong.  It seems like it should be -ge 1 to assign the $ret value to anything that is not 0.  Let me know if I'm missing something here.

+++ plugins/unpack/unrar_file.sh
@@ -14,13 +14,12 @@ else
 	ret=$?
 fi
 
-ret=$?
-if [ $ret -le 1 ] && [ "$5" != '' ] ; then
+if [ $ret -eq 0 ] && [ "$5" != '' ] ; then
 	rm "$5"
 fi
 
 if [ "$6" != '' ] ; then
 	mkdir -p "$3"
 	mv "$6"* "$3"
-	rm -r "$6"
+	[ $? -eq 0 ] && rm -r "$6"
 fi
